requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via
    #   kubernetes
    #   requests
charset-normalizer==2.0.12
    # via requests
google-auth==2.6.0
    # via kubernetes
idna==3.3
    # via requests
kubernetes==v23.3.0
    # via remote-pod-debugger (setup.py)
oauthlib==3.2.0
    # via requests-oauthlib
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
python-dateutil==2.8.2
    # via kubernetes
pyyaml==6.0
    # via kubernetes
requests==2.27.1
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
urllib3==1.26.8
    # via
    #   kubernetes
    #   requests
websocket-client==1.3.1
    # via kubernetes

# The following packages are considered to be unsafe in a requirements file:
# setuptools
